Add test code from java_common
When packing java_common from jitpack.io, only the code in main/java is
packed, (which makes sense) which doesn't include any of the code here.
Moving the test halper code to main/java would mean that junit and
hamcrest would now be in the runtime classpath, which we don't want, so
it seems the least bad option is to make a new library.

Also copied over the blobstore and minio controllers from workspace.
The mysql controller is unused to my knowledge so it was not copied.